anon/requirements.txt
#git+https://github.com/naya1503/pyrogram
pyrogram
TgCrypto
kymang==2.0.2
py-tgcalls
pytz
yt-dlp
youtube-search-python
motor==3.0.0
python-dateutil
dnspython==2.3.0
GitPython
feedparser
speedtest-cli
pykeyboard
bs4
Pillow==8.3.2
pymediainfo
opencv-python-headless
requests
ffmpeg-python
unidecode
attrify
pyromod==1.1
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
emoji
wget
telegraph
heroku3
openai
httpx[http2]
uvloop
httpx
python-dotenv
py-extract
psutil
git+https://github.com/Itz-fork/fake-useragent.git
search_engine_parser
future
pytimeparse
flask
flask_restful
psycopg2-binary
sqlalchemy==1.3.23